+++ src/Widgets/TreeView.cpp
@@ -1006,9 +1006,9 @@ namespace tgui
         if (!m_mouseHover)
             mouseEnteredWidget();
 
-        if ((m_verticalScrollbar->isMouseDown() && m_verticalScrollbar->isMouseDownOnThumb()) || m_verticalScrollbar->isMouseOnWidget(pos))
+        if (m_verticalScrollbar->isMouseDown() || m_verticalScrollbar->isMouseOnWidget(pos))
             m_verticalScrollbar->mouseMoved(pos);
-        else if ((m_horizontalScrollbar->isMouseDown() && m_horizontalScrollbar->isMouseDownOnThumb()) || m_horizontalScrollbar->isMouseOnWidget(pos))
+        else if (m_horizontalScrollbar->isMouseDown() || m_horizontalScrollbar->isMouseOnWidget(pos))
             m_horizontalScrollbar->mouseMoved(pos);
         else
         {